Implement basic bandpass filtering for DK Bongos. Increase microphone threshold.

+++ gc/ogc/pad.h
@@ -158,7 +158,7 @@ u8 PAD_TriggerR(s32 chan);
 u8 PAD_AnalogA(s32 chan);
 u8 PAD_AnalogB(s32 chan);
 
-u8 PAD_BarrelMic(s32 chan);
+s16 PAD_BarrelMic(s32 chan);
 
 #define PAD_SteeringDelta			PAD_StickDeltaX
 #define PAD_Steering				PAD_StickX
